Move base framework classes out into Framework directly, allow Triggers/Cooldowns to only house implementations.
Also small improvement to logging to assist with debugging.
