+++ MagicBots/Overseer/Framework/CooldownAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MagicBots.Overseer.Framework.Cooldowns
+namespace MagicBots.Overseer.Framework
 {
     /**
      * Base attribute for cooldown.
