+++ MagicBots/Overseer/Framework/TriggerAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MagicBots.Overseer.Framework.Triggers
+namespace MagicBots.Overseer.Framework
 {
     /**
      * Base attribute for trigger.
